configure: fix the check for MFX_CODEC_VP9
The data structures for VP9 in mfxvp9.h were wrapped by
MFX_VERSION_NEXT, which means those data structures have never been used
in a public release. Actually MFX_CODEC_VP9 and other VP9 stuffs were
added in mfxstructures.h. In addition, mfxdefs.h was included in
mfxvp9.h, so we may use the check in this patch for MFX_CODEC_VP9

This is in preparation for oneVPL support because mfxvp9.h was removed
from oneVPL [1]

[1]: https://github.com/oneapi-src/oneVPL